theoretically allow per-root watcher selection
This is a refactoring that allows each root to have its own watcher
implementation/"class".

There's no way to actually express that at the moment, so we're
functionally the same as before this diff.

A follow-on diff will allow selecting the watcher.

I've removed the global watcher init/dtor and global watcher instance
type from the code in this diff; we didn't use it and it makes it
more difficult to have a per-root watcher "class".