Allow selecting the watcher
Introduces a `watcher` configuration option that specifies the name of
the watcher that should be used.

The default is `auto` which will run through the list of
`available_watchers` and use whichever is first to successfully
initialize.

If a specific watcher is requested and fails to initialize, watchman
will fall back to using `auto`; this is so that we make a best effort to
start successfully.