Port subscription test to python
In the process, expand the getSubscription API in pywatchman so
that it can track and return per-watch subscription data.

This is important because we can bleed data across tests in our test
harness and because we run a matrix for each test case, there are
2/3 chances to collide and see stale data on the pre-existing client
session object.

Expand the test to positively assert that `defer_vcs` is operating
as we expect.