watchman: improve handling of case insensitive filenames
Summary: we hit an issue where a directory rename that changed only the
case of the dir name got us confused.  It had a bad interaction with
the recently added symlink sanity checking code; the problem manifested
as watchman relatively busily ping-ponging between thinking that `DIR/a`
and `dir/a` existed.

The heart of this issue is similar to the one that we solved with
symlinks; at any point, any path component could have changed case, but
on a case insensitive filesystem we can't ask the system to operate in
case sensitive mode for a given operation.

The resolution is similar to our resolution of the symlink issue; as we
walk down from the root, not only do we check for symlinks, but we now
also check the canonical case of the path component.  If what we want
to open doesn't match the canonical case, we report ENOENT.

This gives us a pretty decent case-sensitive stat operation (modulo
TOCTOU, but it's no worse than our existing symlink checking).

In our stat_path/crawler code, we check for ENOENT/ENOTDIR on case
insensitive filesystems and use that as a signal to look in the parent
of the item that we tried to examine.

This serves two purposes:

1. It allows us to discover the new canonical name by walking the dir
   contents.
2. We can garbage collect any other similarly case-changed items in that
   dir.

A nice side effect of this is that we can remove the relatively complex
lc_files hash stuff; our core data structures can all be case sensitive
and we can leave the insensitive matching to the query engine.  This
means that we can halve the memory used to build our filesystem tree
representation on Mac (nice!) but does make it a bit more expensive to
crawl the filesystem.

So, looking at perf; implementing things strictly as described above
adds 50% to the initial crawl time for one of our large repos (so 150%
total time vs. prior implementation).

The original intent of the code that walks down from the root was to
avoid going out to am expensive remote filesystem if a symlink change
surprised us.  Since that type of change is not common, and because
we're happy to now perform case sensitive compares on the path vs. the
canonical path, we can switch to using a system-dependent means for
obtaining the canonical path from an opened file descriptor.

This switch takes us from a negative perf impact to a positive perf
impact; 50% of the prior implementation for crawling this same large
repo.

Test Plan:

Rebasing either side of a directory case change rename followed by `hg
status` does not show any phantom files.

`make integration` passes.

This little program shows that F_GETPATH works as intended, and also on
mounted drives:

```
$ cat norm.c

void info(const char *path) {
  char *real = realpath(path, NULL);

  printf("Input path: %s\n", path);
  printf("Realpath: %s\n", real);

  int fd = open(path, O_RDONLY);
  if (fd != -1) {
    char buf[MAXPATHLEN];

    buf[0] = 0;
    int res = fcntl(fd, F_GETPATH, buf);

    printf("F_GETPATH: %d %s\n", res, buf);
  }

  printf("\n");
}

int main(int argc, char**argv) {
  int i;

  for (i = 1; i < argc; i++) {
    info(argv[i]);
  }
  return 0;
}
```

```
$ ./norm /Volumes/mercurial/.temporaryitems /Volumes
Input path: /Volumes/mercurial/.temporaryitems
Realpath: /Volumes/mercurial/.TemporaryItems
F_GETPATH: 0 /Volumes/mercurial/.TemporaryItems

Input path: /Volumes
Realpath: /Volumes
F_GETPATH: 0 /Volumes
```