subscriptions: update the ticks when a query yields no results
We've been tracking an issue where we behave poorly when a subscription
doesn't match any files after an extended period of changes.

In that situation, if the root is currently on tick 400 and the
subscription last matched at tick 10, every observed change will
cause a walk over all the files walking from most recent to eldest,
until it finds any that had tick<=10.  This can take hundreds of
milliseconds in a busy tree.

This diff causes us to update the subscription clockspec in the case
that it yielded no results.

Test Plan: passes `make integration` in this and
https://github.com/facebook/watchman/pull/187