fix subscription/initial crawl race condition
It's possible for the subscription request to complete (with an empty
initial set) if it is placed immediately after initiating a watch but
before the crawl is complete.  This can happen because subscriptions
don't use query synchronization.

This race happened to be won/lost consistently on the Travis CI infra
and caused problems because the test wanted to observe the initial
file results for a freshly created directory structure before it
would proceed and make changes.  That effectively caused a deadlock
because we wouldn't send any of the changes in that view until a
subsequent filesystem change was observed.

The simple resolution is to increment the ticks prior to recording
the file generated by either the initial crawl or subsequent recrawls