refactor: add watchman_user_client
This adds watchman_user_client to represent client connections made
to the watchman per-user service.

This is distinct from the base watchman_client struct which we'd
like to use for connections to the devfsevents daemon connections.

Factoring it out means that we don't need to pull in various root
related modules when building the devfsevents daemon itself.

The implementation of this is not the most elegant; we're effectively
adding in some C++isms but in raw C and its a bit gross.  I wanted to
limit the scope of the changes in this diff to get the ball rolling on
the actual daemon code.  I will revisit the listener/client mechanic in
a follow on diff and see what we can do to make it nicer.