Cancel subscriptions when we cancel a root
It turns out that we were not fully cleaning up associations from a
client subscription to cancelled roots.  This could cause test_age_watch
to fail intermittently because the root pointer was later re-used for a
different watch and we'd notice that the long lived client session from
a deleted watch matched the root; that would block us from aging our the
new incarnation of the root.

When the final ref on a root is about to go, let's walk over the client
list and generate a cancellation subscription notification and remove
the association.