critbit: add function to find longest match
Summary: Locates the external node with the longest common prefix with
an input string.

We can't directly compare the key until we've arrived at an external
node so this function remembers the prior branch point.  If we land on
an external node that is logically after the input string, we backtrack
to that branch point and walk out to the alternate external node.

This seems to be sufficient for the tests I've included.  I'm worried
that I'm missing something important though.