fix longstanding `watchman -p --server-encoding=json` bug
When using the watchman CLI in persistent mode, we'd hit an error
identifying the PDU.  The workaround was to specify json encoding.

I finally got around to looking into this and the cause is quite simple:
we're hitting an EAGAIN error and not setting a very good error message
(we skip formatting it because we consider it to be a frequent and
uninteresting condition).

We get EAGAIN because the bser PDU reader unconditionally enables and
disables non-blocking at certain phases, so even though we started out
in blocking mode, after reading the initial response from the server we
end up in non-blocking and fall over.

The resolution is simply to fixup the passthru mode of the CLI to reset
us back to blocking mode just prior to decoding the next PDU header.