add w_root_lock_with_timeout
This is helpful when we wish to set an upper bound on how long we wish
to wait for watchman to become ready when running queries.

I've introduced a new lock_timeout field to the query definition.  It
defaults to 1 minute.

If the lock cannot be acquired within the specified lock_timeout, then
we will log the reason to the log file and include the current
lock_reason in the error string that we return in the error response to
the client making the query.