add w_perf_t for recording performance
This rethinks the earlier slow command logging and formalizes the
approach so that we can maintain a context and associate metadata with
it.  If a region of code takes "too long" then we can log some
interesting performance and timing information.

In addition to logging a PERF line to the log file, we allow for a
`perf_logger_command` to configured in the `/etc/watchman.json`
configuration file.  If set, it is a program that will receive a list of
PERF json blobs as arguments.

The intent is that that command can be used to send perf samples to some
site-specific aggregator.