Add concept of session owner, restrict command execution
Refs: https://github.com/facebook/watchman/issues/222

This adds an interface to our stream layer that allows us to check
whether the peer matches the owner of our process.  We do not
currently implement this for Windows.

We record this flag in the client structure.

The command dispatcher will by default refuse to allow any commands to
run if the peer is not the owner.

A command can be registered with the `CMD_ALLOW_ANY_USER` to whitelist
that command from this blanket dispatcher check.

The model we're using to decide what is white listed is this:

* If you are not the owner, you can inspect any of the current
  state in the watchman instance.
* You may not change that state; initiating new watches, deleting
  watches or registering triggers are prohibited

`watch` and `watch-project` are white listed because they are an
important part of initializing a client; well behaved clients will
issue `watch-project` to resolve the watched path and directory
offset they should use for their session.

For that reason, we add a check for client==owner when we resolve a
root; if the `auto_create` flag is true and you are not the owner,
we turn the flag off.  We'll also expand the error message to provide
more context.

Test Plan: included a test case, but also manually:

```
$ sudo watchman -U /opt/facebook/watchman/var/run/watchman/wez-state/sock --no-spawn watch list
{
    "version": "4.5.0",
    "error": "unable to resolve root list: path \"list\" must be absolute (this may be because you are not the process owner)"
}
$ sudo watchman -U /opt/facebook/watchman/var/run/watchman/wez-state/sock --no-spawn watch-list
{
    "version": "4.5.0",
    "roots": [
        "/Users/wez/fbsource",
        "/Users/wez/fb/watchman"
    ]
}
$ sudo watchman -U /opt/facebook/watchman/var/run/watchman/wez-state/sock --no-spawn watch-del /Users/wez/fbsource
{
    "version": "4.5.0",
    "error": "you must be the process owner to execute 'watch-del'"
}
$ sudo watchman -U /opt/facebook/watchman/var/run/watchman/wez-state/sock --no-spawn watch /Users/wez/fbsource
{
    "version": "4.5.0",
    "watch": "/Users/wez/fbsource",
    "watcher": "fsevents"
}
```