Support for symbolic link target handling
Summary: Add a field for symlink_target in watchman file.
Add a query field to return the symlink_target for the file.

When symbolic link changes, update the target metadata.

Test Plan: Added tests for symlink. Some of these tests are
not run on Darwin and NT, because of platform specific behavior.

1. test for basic symlink targets.
2. invalid symlinks.
3. deletion of symlinks.
4. transition of symlink to regular file or directory

Reviewers: wez

Subscribers:

Tasks:

Blame Revision: