libart: remove some node casts
Summary: This is a stylistic change that replaces a number of inline
node type casts with a union (and a const flavor of that union).

This helps to reduce the amount of code we have to read. In addition
removing casts allows the compiler to catch more mistakes, especially
during refactoring.

The union trick was already being used in a couple of places; the new
union type introduced here is a little different and (subjectively)
nicer because the old type used the node enum values to differential
which element was being used whereas this implementation now uses the
node size, making it easier to visually verify that the correct type is
being used.