tests: separate out code that initializes watchman files into a mixin
Setting up the files beforehand is generally useful, but unfortunately it
doesn't catch issues with state directory creation or permissions. So we need
to have a separate kind of instance, meant specifically to test state dirs.
Separating this code out into a mixin will allow us to also plug in a separate
mixin later.