disable use of bulkstat on older Darwin kernels
Summary: I've seen this problem once in the past couple of months
and we worked around it by setting the kill switch option.

I remembered this just now while prepping for the 4.6 release; I don't
want to break the user that worked around this (the update will blow
away their modified watchman.json file), so let's tuck in this check
to set the default.

Test Plan: I added a W_LOG_FATAL to log the values of `safe`, `maj`,
`min`, `patch` and visually verified that they match my system and that
`safe` is `true`, indicating that we are safe to use bulk stat.