avoid a race condition during crawling
Summary: I suspect that the issue observed by @simpkins was due to a
race during the initial crawl.

The explanation of the race is inline in comments in the code.

Test Plan: I don't think there is a reliable way to trigger this and/or
prove it short of rolling it out.  The integration tests continue to
pass as they did before.