avoid a race condition during crawling
Summary: I suspect that the issue observed by @simpkins was due to a
race during the initial crawl.

The explanation is inline.

Test Plan: I don't think there is a reliable way to trigger this short
of rolling it out.  The integration tests continue to pass as they did
before.