fixup prefix iteration
Summary: This was a bit of a surprise and caused some incorrect
behavior.  I added the logging first to gain some insight and found
that we were overly aggressive at pruning in delete_kids.

I tracked this down to the partial prefix matching code in the prefix
iterator; one of the cases is that the minimum node is located and used
as the basis for a vanilla recursive iteration.

So for example, we would be searching for `/foo/bar/baz` prefixes and
the tree would use `/foo/bar` as the root of a recursive iteration and
emit all the leaves under `/foo/bar` rather than just those that were
logically under `/foo/bar/baz`.

This condition triggers in `tests/integration/age.php` where we stress
test the node aging code and repeatedly create and remove nodes with
similar and same names.   The `w_assert` in the iterator callback
in `pending.c` triggers in non-prod builds in this case.

To address this, the prefix iteration needs to shim in a callback to
pre-check the leaves before they are passed to the user-supplied
callback, and that is what the bulk of the functional changes do in
this diff.

Test Plan: `make integration` would fail once I added the assert but
before I added the iterator callback shim.  After adding the shim,
all tests pass.

Capture the debug logs for updating a big repo via:

```
watchman --server-encoding=json -p log-level debug | egrep
'(delete_kid|maybe_prune|SKIP)' > /data/users/wez/blarg.log
```

and inspect and sanity check the output.

Reviewers: sid0

Subscribers:

Tasks: #11693500

Blame Revision: