speculative fix for some fishy state-leave behavior
Summary: This has been hard to nail down, but we've seen some weird
crashes every so often where we see events like this in the log:

```
2016-06-10T21:34:06,035: [client:stm=0x63c5c0] implicitly vacating state hg.update on /data/users/wez/FOO due to client disconnect
2016-06-10T21:34:24,367: [client:stm=0x63c5c0] send_error_response: ["state-leave", "/data/users/wez/FOO", {"name": "hg.update", "metadata": {"rev": "hash1 hash2", "distance": 123, "status": "ok", "partial": false}}] failed: state hg.update is not asserted
```

this doesn't make sense because the the first line happens 20 seconds
before the second, but can't possibly run before the latter.

I haven't found a way to reproduce this even intermittently, let alone
reliably, so this diff adds some belt and suspenders stuff around the
cleanup of this so that we can hopefully get a better read on this
if/when it happens in the future.

Test Plan: make integration + CI