add site-specific spawn command
Summary: in some situations, the system administrator may desire to have
control over the context in which the watchman service is started.

For example, if may be desirable to force the service to spawn a
system service that was automatically created for the current user
rather than just having watchman fork off a child.

For this situation we now provide a `spawn_watchman_service`
configuration key that is available in `/etc/watchman.json`.
If this string is set, when the watchman CLI decides that it is
appropriate to spawn the watchman service, rather than using the
default spawning action, it will instead invoke `spawn_watchman_service`
and forward all daemon appropriate watchman arguments to it.

The value of `spawn_watchman_service` will be used as `argv[0]` when
launching the spawner, so it must either be the full path to the
spawning program, or be something that posix_spawn is able to resolve
from the PATH of the current process.

If the spawner returns with a non-zero exit status, the watchman command
that invoked it will fail and report the problem to the user.

The stdout and stderr of the spawner program are redirected to the
watchman log file.