eliminate a string copy in the fsevents callback
This removes a stack buffer and a copy.  We used to require that we have
a NUL terminated string for some flavor of the old ignore list
processing, but the new art based code doesn't have this restriction.

Optimize slightly but avoiding a lock/signal if we ignored everything
that was in the notification callback, and by deferring the logging
to the thread that consumes these.