add integration test for the resync
This asserts that the resync leaves us with a correct state,
and turns on this feature by default.

I inspected the log files and can see that we receive the cookie
notifications for a prior cookie again after we've set up the
replacement stream.  It looks good, but we can't easily make assertions
on the log file contents in the test without a high chance of flakiness.