Take another stab at the zombie problem
Summary: we've seen some zombies hanging out under watchman.  Manually
running `pkill -CHLD <pid-of-reaper-thread>` woke up the reaper to
collect them, so that implies that we have another race.

From looking at the code a bit harder, the race condition is likely that
the threads that we spawn from the state file are started before we mask
off SIGCHLD.

Because posix and threads are awesome, this means that the kernel can
choose to deliver SIGCHLD to an arbitrary thread that hasn't explicitly
blocked it.  This results in an EINTR somewhere, but not in the child
getting reaped.

This diff masks off SIGCHLD early, and arranges for the reaper to start
up before we spawn any threads, just to be sure that we have something
hanging around and ready to reap any children that might be start by
triggers loaded from the state file.