avoid spuriously updating directories on darwin
Summary: the pantsbuild folks reported that they receive spurious
notifications for a parent directory when touching a child directory.

In tracing through the code, this is due to the logic in this diff
that desires to trigger a stat on the parent directory.  This behavior
was introduced in c6dbda20726d58f578bbca290c6a739c4e76d751 to ensure
that we notice mtime updates for the parent.

In this situation, since the event that triggered for the child node
had VIA_NOTIFY set, we will propagate this to the stat_path call when
looking at the parent and this will cause it to unconditionally mark
the directory node as changed (because the criteria is `via_notify ||
did_file_change()`).

I think the solution here is just to zero out the flags that we pass
down and that is what I've done here in this diff.  The existing
twodeep.php integration test continues to pass for me on the mac so this
doesn't obviously break things.  The diff that introduced this behavior
was Linux specific though, so I'm interested to see what our Linux CI
makes of this.

Closes https://github.com/facebook/watchman/issues/305
Closes https://github.com/facebook/watchman/issues/307