Reduce memory utilization around dir names
We've traditionally maintained a map of full dir name -> dir structure.
As the volume and depth of a watched tree increases, this results in
a lot of duplicate path prefixes maintained in memory, bloating the
watchman process.

This diff trades the O(1) lookup time complexity for O(depth-of-path-in-root)
for a reduced memory footprint; we're now storing just a single path
component in each directory and file.  This means that we have to
introduce some helper functions to compute the effective directory
path on demand.

This translates into about 40% memory savings (but see note below)
and should be more on OSX.  I've only done some light testing with
this because my local repos are more sparse than is representative; I
will look at this when I'm plugged into the ethernet.

Note that inotify and kqueue watches maintain additional maps keyed by
dirname; this diff does not change that, and there is potentially
more to be saved by looking at that in more detail.