introduce a write lock holder
This is the first step towards removing the recursive giant lock that we
use in the w_root_t.

This diff introduces a writeable lock holder type that we can pass
around instead of the w_root_t.  This allows helper functions to
explicitly advertize the locking that they expect to be in place.

The locking functions have changed to:

```
w_root_lock(&root, "reason", &lock);
do_something(&lock);
root = w_root_unlock(&lock);
```

The `root` variable is NULL'd out for the duration that the lock is
held; this is sort of like doing the follow C++-style assignment:

```
lock = w_root_lock(std::move(root), "reason");
```

this is so that the compiler can catch references to `root->` in this
first sweep.

I've tried to contain this to just the changes needed to support this
new locking protocol, although I did take the opportunity to change a
couple of functions to explicitly require a `write_locked_watchman_root`
parameter to demonstrate what this looks like.

I'll follow up with diff(s) to cut most helper functions over to
accepting a `write_locked_watchman_root`, then we should be in a better
position to remove the recursive locking.  With that removed, we can
then lock at reader-writer locking.